requirements.txt
torch
torchvision
pyctcdecode
numpy
tqdm
tensorboard
matplotlib
wandb
jiwer


pandas
speechbrain~=0.5.9
torchaudio~=0.9.1
datasets
torch_audiomentations
editdistance
torch_optimizer
wget